tools/src/pyproject.toml
[build-system]
requires = ["flit_core==3.8.0"]
build-backend = "flit_core.buildapi"

[project]
name = "scie-pants"
description = "Support tools for scie-pants."
dynamic = ["version"]
requires-python = ">=3.8,<3.12"
dependencies = [
    "ansicolors==1.1.8",
    "conscript==0.1.5",
    "tomlkit==0.11.6",
]

[[project.authors]]
name = "The Pants Developers"
email = "pantsbuild@gmail.com"

[project.license]
file = "LICENSE"

[project.scripts]
bootstrap-tools = "scie_pants.bootstrap_tools:main"
install-pants = "scie_pants.install_pants:main"
update-scie-pants = "scie_pants.update_scie_pants:main"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = "colors"
ignore_missing_imports = true
